% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gwasapi.R
\name{check_gwasapi}
\alias{check_gwasapi}
\title{Check if the GWAS Catalog is live}
\usage{
check_gwasapi()
}
\value{
Character string.
}
\description{
Check if the GWAS Catalog is live
}
\examples{
check_gwasapi()
}
