% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gwascat.R
\name{check_gwascat}
\alias{check_gwascat}
\title{Check if the GWAS Catalog is live}
\usage{
check_gwascat(url = gwascat_url())
}
\arguments{
\item{url}{URL for the GWAS Catalog}
}
\value{
Character string.
}
\description{
Check if the GWAS Catalog is live
}
\examples{
check_gwascat()
}
