% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_associations.R
\name{get_asso}
\alias{get_asso}
\title{Get GWAS associations for a given region}
\usage{
get_asso(
  chr,
  bp_lower = NULL,
  bp_upper = NULL,
  study = NULL,
  p_lower = NULL,
  p_upper = NULL,
  start = NULL,
  size = NULL
)
}
\arguments{
\item{chr}{chromosome number}

\item{bp_lower}{Lower endpoint of basepairs initerval}

\item{bp_upper}{Upper endpoint of basepairs initerval}

\item{study}{Restrict to a particular study}

\item{p_lower}{Lower bound on p-values}

\item{p_upper}{Upper bound of p-values}

\item{start}{First record to retrieve (starting at 0)}

\item{size}{Maximum number of results to retrieve}
}
\value{
Data frame of associations
}
\description{
Get GWAS associations for a given region
}
\examples{
result <- get_asso(chr=19, bp_lower=19200000, bp_upper=19300000)
}
\seealso{
\code{\link[=get_variant]{get_variant()}}, \code{\link[=get_trait_asso]{get_trait_asso()}}
}
