% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_associations.R
\name{get_trait_asso}
\alias{get_trait_asso}
\title{Get GWAS associations for a particular trait}
\usage{
get_trait_asso(
  trait = NULL,
  study = NULL,
  p_lower = NULL,
  p_upper = NULL,
  start = NULL,
  size = NULL
)
}
\arguments{
\item{trait}{Restrict to a particular study}

\item{study}{Restrict to a particular study}

\item{p_lower}{Lower bound on p-values}

\item{p_upper}{Upper bound of p-values}

\item{start}{First record to retrieve (starting at 0)}

\item{size}{Maximum number of results to retrieve}
}
\value{
Data frame of associations
}
\description{
Get GWAS associations for a particular trait
}
\examples{
\donttest{result <- get_trait_asso("EFO_0001360", p_upper=1e-10)}
}
\seealso{
\code{\link[=get_variant]{get_variant()}}, \code{\link[=get_asso]{get_asso()}}
}
