% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_associations.R
\name{get_variant}
\alias{get_variant}
\title{Get associations for a given variant}
\usage{
get_variant(
  rsnum,
  chr = NULL,
  p_lower = NULL,
  p_upper = NULL,
  study = NULL,
  start = NULL,
  size = NULL
)
}
\arguments{
\item{rsnum}{RS number for a variant}

\item{chr}{chromosome number if known}

\item{p_lower}{Lower bound on p-values}

\item{p_upper}{Upper bound of p-values}

\item{study}{Restrict to a particular study}

\item{start}{First record to retrieve (starting at 0)}

\item{size}{Maximum number of results to retrieve}
}
\value{
Data frame with associations as rows
}
\description{
Get all associations for a given variant
}
\examples{
# get associations for a given variant
results <- get_variant("rs2228603")
# use information about the chromosome it's on
results <- get_variant("rs2228603", 19)
# get the next 20 results
next20 <- get_variant("rs2228603", 19, start=20)
# get 100 results rather than just 20
first100 <- get_variant("rs2228603", 19, size=100)
# return just the associations with P < 1e-8
top_results <- get_variant("rs2228603", 19, p_upper=1e-8)

}
\seealso{
\code{\link[=get_asso]{get_asso()}}, \code{\link[=get_trait_asso]{get_trait_asso()}}
}
