% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASapi-package.R
\docType{package}
\name{GWASapi-package}
\alias{GWASapi}
\alias{GWASapi-package}
\title{GWASapi: NHGRI-EBI GWAS Catalog of Summary Statistics}
\description{
Tools for connecting to the NHGRI-EBI API for GWAS summary statistics.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/GWASapi/GWASapi.html}{GWASapi user guide}
}
}

\section{API documentation}{

\itemize{
\item \href{https://www.ebi.ac.uk/gwas/summary-statistics/docs/}{NHGRI-EBI GWAS Catalog API documentation}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rqtl/GWASapi}
  \item Report bugs at \url{https://github.com/rqtl/GWASapi/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
