% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasapi_url.R
\name{gwasapi_url}
\alias{gwasapi_url}
\title{The URL for the GWAS Catalog API}
\usage{
gwasapi_url()
}
\value{
The URL, as character string
}
\description{
The current URL for the GWAS Catalog API
}
\examples{
gwasapi_url()

}
