% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAScat-package.R
\docType{package}
\name{GWAScat-package}
\alias{GWAScat}
\alias{GWAScat-package}
\title{GWAScat: NHGRI-EBI GWAS Catalog of Summary Statistics}
\description{
Tools for connecting to the NHGRI-EBI API for GWAS summary statistics.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/GWAScat/GWAScat.html}{GWAScat user guide}
}
}

\section{API documentation}{

\itemize{
\item \href{https://www.ebi.ac.uk/gwas/summary-statistics/docs/}{NHGRI-EBI GWAS Catalog API documentation}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rqtl/GWAScat}
  \item Report bugs at \url{https://github.com/rqtl/GWAScat/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (0000-0002-4914-6671)

}
\keyword{internal}
