% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascat_url.R
\name{gwascat_oldurl}
\alias{gwascat_oldurl}
\title{The URL for the GWAS Catalog API}
\usage{
gwascat_oldurl()
}
\value{
The URL, as character string
}
\description{
The current URL for the GWAS Catalog API
}
\examples{
gwascat_url()

}
