% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_study}
\alias{info_study}
\title{Information about a study}
\usage{
info_study(study, url = gwascat_oldurl())
}
\arguments{
\item{study}{Study identifieer}

\item{url}{The url for the API (actually the old API)}
}
\value{
List of meta-data
}
\description{
Information about a particular study
}
