% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stuff.R
\name{list_chromosomes}
\alias{list_chromosomes}
\title{List chromosomes}
\usage{
list_chromosomes(url = gwascat_url())
}
\arguments{
\item{url}{URL for GWAS catalog API}
}
\value{
Vector of character strings with chromosome numbers
}
\description{
List chromosomes
}
\examples{
chr <- list_chromosomes()

}
