% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stuff.R
\name{list_studies}
\alias{list_studies}
\title{List studies}
\usage{
list_studies(url = gwascat_url(), start = NULL, size = NULL)
}
\arguments{
\item{url}{The URL of the GWAS Catalog API}

\item{start}{First record to retrieve (starting at 0)}

\item{size}{Maximum number of results to retrieve}
}
\value{
Vector of study accessions
}
\description{
List studies
}
\examples{
list_studies() # returns 20 studies
list_studies(start=20) # returns the next 20 studies
list_studies(size=100) # returns 100 studies
}
