% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stuff.R
\name{list_traits}
\alias{list_traits}
\title{List traits}
\usage{
list_traits(url = gwascat_url(), start = NULL, size = NULL)
}
\arguments{
\item{url}{The URL of the GWAS Catalog API}

\item{start}{First record to retrieve (starting at 0)}

\item{size}{Maximum number of results to retrieve}
}
\value{
Vector of trait IDs
}
\description{
List traits
}
\examples{
first20 <- list_traits() # first 20 traits
next20 <- list_traits(start=20) # the next 20 traits
first100 <- list_traits(size=100) # returns 100 traits
}
\seealso{
\code{\link[=list_studies]{list_studies()}}, \code{\link[=list_chr]{list_chr()}}
}
