% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_associations.R
\name{get_variant}
\alias{get_variant}
\title{Get associations for a given variant}
\usage{
get_variant(rsnum, chr = NULL, p_lower = NULL, p_upper = NULL,
  url = gwascat_url())
}
\arguments{
\item{rsnum}{RS number for a variant}

\item{chr}{chromosome number if known}

\item{p_lower}{Lower bound on p-values}

\item{p_upper}{Upper bound of p-values}
}
\value{
Data frame with associations as rows
}
\description{
Get all associations for a given variant
}
\examples{
get_variant("rs2228603")
)get_variant("rs2228603", 19)
}
