% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{info_study}
\alias{info_study}
\title{Information about a study}
\usage{
info_study(study)
}
\arguments{
\item{study}{Study identifieer}
}
\value{
List of meta-data
}
\description{
Information about a particular study
}
