% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stuff.R
\name{list_chr}
\alias{list_chr}
\title{List chromosomes}
\usage{
list_chr(url = gwascat_url())
}
\arguments{
\item{url}{URL for GWAS catalog API}
}
\value{
Vector of chromosome numbers (as integers)
}
\description{
List chromosomes
}
\examples{
chr <- list_chr()

}
