% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stuff.R
\name{list_studies}
\alias{list_studies}
\title{List studies}
\usage{
list_studies(url = gwascat_url(), start = NULL, size = NULL,
  trait = NULL)
}
\arguments{
\item{url}{The URL of the GWAS Catalog API}

\item{start}{First record to retrieve (starting at 0)}

\item{size}{Maximum number of results to retrieve}

\item{trait}{If provided, restrict search to studies that included the specified trait.}
}
\value{
Vector of study accessions
}
\description{
List studies
}
\examples{
first20 <- list_studies() # returns 20 studies
next20 <- list_studies(start=20) # returns the next 20 studies
first100 <- list_studies(size=100) # returns 100 studies
with_trait <- list_studies(trait="EFO_0001360")
}
\seealso{
\code{\link[=list_chr]{list_chr()}}, \code{\link[=list_traits]{list_traits()}}
}
